#84 feat: API 중복 호출 방지 및 쿼리 무효화 로직 추가
