+++ frontend/src/features/goals/mutations/useAddSavingGoal.ts
@@ -1,13 +1,17 @@
-import { useMutation } from '@tanstack/react-query';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { requestAddSavingGoal, SavingGoalRequest } from '@/features/goals/api/requestAddSavingGoal';
 import { useNavigate } from 'react-router-dom';
 
 export const useAddSavingGoal = () => {
   const navigate = useNavigate();
+  const queryClient = useQueryClient();
 
   return useMutation({
     mutationFn: (data: SavingGoalRequest) => requestAddSavingGoal(data),
     onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: ['remainingBudget'] });
+      queryClient.invalidateQueries({ queryKey: ['dailyExpenses'] });
+      queryClient.invalidateQueries({ queryKey: ['weeklyExpenseStatus'] });
       navigate('/main');
     },
     onError: (error: any) => {
